Create README.md
The original project was committed via my local Git to this remote git 15+ times. No idea why the commits do not show. I followed the same process with the 
first part of the lab and you can see 13+ commits. This one only has two commits and required me to manually add my code to the repo.
Not sure why it didn't add the first time around. Re-checked it this morning and no code was present. 
I tried to push it via Git and it didn't work. I added it in the GUI, I normally do not do that.